Stub for computer batting; fixed movement multiplier
The computer correctly calculates the swing adjust radius, then adjusts
the radius within that on each turn.

There was a bug where the pitch speed multiplier and movement
multiplier were accidentally switched; this is fixed now.  Your pitch
will have less movement on it the farther away from your sweet spot it
is.

Also stubbed out a particle system that could be used for the ball
trail.