Implemented Result Debug View
The ResultDebug View is now correctly working; it accepts inputs for
the characters and location, and renders the subsequent result as a
moving ball.

The BallHit model was also created, to store the ball’s end point and
the time it took to get there.

Still to be determined:
* How do we represent the Vector2 for how far the ball travels?  What’s
the scale?